Rename protocol.abc to protocol.base_io
This rename is necessary as it prepares the file to be used for other
things than just abstract classes. These other uses should be present in
the same file, and will include major changes to the file, this commit
just prepares for that.
